Allow provider assignment from `@DeferredApi` methods.
