+++ tools/lint/src/main/kotlin/DeferredApiDetector.kt
@@ -23,18 +23,10 @@ import com.android.tools.lint.detector.api.JavaContext
 import com.android.tools.lint.detector.api.Scope
 import com.android.tools.lint.detector.api.Severity
 import com.android.tools.lint.detector.api.SourceCodeScanner
-import com.intellij.lang.jvm.JvmModifier
-import com.intellij.psi.PsiClassType
 import com.intellij.psi.PsiMethod
 import org.jetbrains.uast.UAnnotation
-import org.jetbrains.uast.UAnonymousClass
 import org.jetbrains.uast.UCallExpression
 import org.jetbrains.uast.UElement
-import org.jetbrains.uast.ULambdaExpression
-import org.jetbrains.uast.UMethod
-import org.jetbrains.uast.getParentOfType
-
-internal const val ANNOTATION = "com.google.firebase.annotations.DeferredApi"
 
 class DeferredApiDetector : Detector(), SourceCodeScanner {
     override fun applicableAnnotations(): List<String> = listOf(ANNOTATION)
@@ -69,64 +61,6 @@ class DeferredApiDetector : Detector(), SourceCodeScanner {
                             includeReceiver = false,
                             includeArguments = true),
                     "${method.name} is only safe to call in the context of a Deferred<T> dependency.")
-    }
-    private fun hasDeferredApiAnnotation(context: JavaContext, methodCall: UElement): Boolean {
-        lambdaMethod(methodCall)?.let {
-            return hasDeferredApiAnnotation(context, it)
-        }
-
-        val method = methodCall.getParentOfType<UElement>(
-                UMethod::class.java, true,
-                UAnonymousClass::class.java, ULambdaExpression::class.java
-        ) as? PsiMethod
-        return hasDeferredApiAnnotation(context, method)
-    }
-
-    private fun lambdaMethod(element: UElement): PsiMethod? {
-        val lambda = element.getParentOfType<ULambdaExpression>(
-                ULambdaExpression::class.java, true, UMethod::class.java, UAnonymousClass::class.java)
-                ?: return null
-
-        val type = lambda.functionalInterfaceType
-        if (type is PsiClassType) {
-            val resolved = type.resolve()
-            if (resolved != null) {
-                return resolved.allMethods.firstOrNull { it.hasModifier(JvmModifier.ABSTRACT) }
-            }
-        }
-        return null
-    }
-
-    private fun hasDeferredApiAnnotation(context: JavaContext, calledMethod: PsiMethod?): Boolean {
-        var method = calledMethod ?: return false
-
-        while (true) {
-            for (annotation in method.modifierList.annotations) {
-                annotation.qualifiedName?.let {
-                    if (it == ANNOTATION) {
-                        return@hasDeferredApiAnnotation true
-                    }
-                }
-            }
-            method = context.evaluator.getSuperMethod(method) ?: break
-        }
-
-        var cls = method.containingClass ?: return false
-
-        while (true) {
-            val modifierList = cls.modifierList
-            if (modifierList != null) {
-                for (annotation in modifierList.annotations) {
-                    annotation.qualifiedName?.let {
-                        if (it == ANNOTATION) {
-                            return@hasDeferredApiAnnotation true
-                        }
-                    }
-                }
-            }
-            cls = cls.superClass ?: break
-        }
-        return false
     }
 
     companion object {
